// src/enums/cardValue.ts
export enum UnoValue {
    Zero = '0',
    One = '1',
    Two = '2',
    Three = '3',
    Four = '4',
    Five = '5',
    Six = '6',
    Seven = '7',
    Eight = '8',
    Nine = '9',
    PlusTwo = '+2',
    Reverse = 'Reverse',
    Pass = 'Pass',
    PlusFour = '+4',
    ChangeColor = 'ChangeColor'
}

export enum SkyjoValue {
    MinusTwo = '-2',
    MinusOne = '-1',
    Zero = '0',
    One = '1',
    Two = '2',
    Three = '3',
    Four = '4',
    Five = '5',
    Six = '6',
    Seven = '7',
    Eight = '8',
    Nine = '9',
    Ten = '10',
    Eleven = '11',
    Twelve = '12'
}